requirements.txt
google-adk[extensions]>=1.0.0
streamlit>=1.40.0
pandas>=2.2.0
openpyxl>=3.1.0
matplotlib>=3.9.0
plotly>=5.24.0
scipy>=1.14.0
sqlalchemy>=2.0.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.32.0
python-dotenv>=1.0.0
fpdf2>=2.8.0
kaleido>=0.2.1
sseclient-py>=1.8.0
